started working on making syscalls not rely on undefined behavior
ABI is unchanged

+++ shared/sys/sys_public.h
@@ -91,7 +91,7 @@ void	Sys_Init (void);
 // general development dll loading for virtual machine testing
 typedef void *GetGameAPIProc( void  *);
 typedef intptr_t QDECL VMMainProc( int, ... );
-typedef intptr_t QDECL SystemCallProc( intptr_t, ... );
+typedef intptr_t QDECL SystemCallProc( int, ... );
 typedef void * QDECL GetModuleAPIProc( int, ... );
 
 void	*Sys_LoadSPGameDll( const char *name, GetGameAPIProc **GetGameAPI );
